Refactor color variables and enhance accessibility; add new constants for process and footer data . PR resolved
