+++ src/app/components/FaqSection.tsx
@@ -2,6 +2,7 @@
 
 import React, { useState } from "react";
 import { ChevronDown, ChevronUp } from "lucide-react";
+import { initialFaqData , additionalFaqData } from "../constants/faqData";
 
 interface FAQItemProps {
   question: string;
@@ -40,59 +41,13 @@ const FAQSection = () => {
   const [openIndex, setOpenIndex] = useState<number | null>(null);
   const [showAllFAQs, setShowAllFAQs] = useState(false);
 
-  const initialFaqData = [
-    {
-      question: "1. What services does Documitra offer?",
-      answer:
-        "Documitra provides comprehensive visa and passport assistance services, including document processing, application support, and expert guidance throughout your journey.",
-    },
-    {
-      question: "2. How does the Documitra process work?",
-      answer:
-        "Our streamlined process involves three simple steps: document submission, verification, and processing. Our experts guide you through each phase to ensure a smooth experience.",
-    },
-    {
-      question: "3. Is Documitra a government agency?",
-      answer:
-        "No, Documitra is an independent service provider that specializes in visa and passport assistance. We work alongside government agencies to facilitate your application process.",
-    },
-    {
-      question: "4. Is my personal information secure with Documitra?",
-      answer:
-        "Yes, we maintain the highest standards of data security. All personal information is encrypted and protected according to international privacy regulations.",
-    },
-    {
-      question:
-        "5. How long does it take to process a visa or passport application with Documitra?",
-      answer:
-        "Processing times vary depending on the type of application and the government's processing speed. We strive to expedite the process, but specific timelines may differ based on individual circumstances and government requirements.",
-    },
-  ];
-
-  const additionalFaqData = [
-    {
-      question: "6. What documents do I need to provide?",
-      answer:
-        "Required documents vary by application type. We'll provide a detailed checklist based on your specific needs and guide you through the collection process.",
-    },
-    {
-      question: "7. Do you offer expedited services?",
-      answer:
-        "Yes, we offer expedited processing for urgent applications, subject to availability and additional fees. Contact us to learn more about rush service options.",
-    },
-    {
-      question: "8. What countries do you service?",
-      answer:
-        "We assist with visa and passport applications for most major countries. Contact our team to confirm availability for your specific destination.",
-    },
-  ];
 
   const displayedFAQs = showAllFAQs
     ? [...initialFaqData, ...additionalFaqData]
     : initialFaqData;
 
   return (
-    <div className="w-full bg-bgBlue py-16">
+    <div className="w-full bg-secondary py-16">
       <div className=" px-[10vw] mx-auto  lg: py-16 flex flex-row">
         <div className=" my-auto mr-10 w-1/3">
           <h2 className="text-5xl font-bold mb-10">
@@ -104,7 +59,7 @@ const FAQSection = () => {
         </div>
         <div className="l-5 m w-2/3">
           <div className="space-y-2">
-            {displayedFAQs.map((faq, index) => (
+            {displayedFAQs?.map((faq, index) => (
               <FAQItem
                 key={index}
                 question={faq.question}
