+++ src/app/components/GetStartedButton.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 
 const GetStartedButton: React.FC = () => {
   return (
-    <button className="px-16 py-3 bg-primary text-white rounded-full hover:bg-primary transition-colors">
+    <button className="px-16 py-3 bg-primary text-white rounded-full hover:bg-primary transition-colors  " 
+    aria-label="Get Started">
       Get Started
     </button>
   );
