+++ src/app/components/ProcesSection.tsx
@@ -1,9 +1,6 @@
 import React from "react";
 import Image from "next/image";
-import step1 from "../../../public/assets/images/Home/step1.svg";
-import step2 from "../../../public/assets/images/Home/step2.svg";
-import step3 from "../../../public/assets/images/Home/step3.svg";
-import step4 from "../../../public/assets/images/Home/step4.svg";
+import { steps } from "../constants/processDAta";
 
 interface ProcessStepProps {
   imageSrc: string;
@@ -42,28 +39,7 @@ const ProcessStep = ({ imageSrc, title, isLast = false }: ProcessStepProps) => (
 );
 
 const ProcessSection = () => {
-  const steps = [
-    {
-      title: "Step 01",
-      imageSrc: step1,
-      isLast: false,
-    },
-    {
-      title: "Step 02",
-      imageSrc: step2,
-      isLast: false,
-    },
-    {
-      title: "Step 03",
-      imageSrc: step3,
-      isLast: false,
-    },
-    {
-      title: "Get Delivered at Your doorstep",
-      imageSrc: step4,
-      isLast: true,
-    },
-  ];
+ 
 
   return (
     <div className="bg-gray-50 py-16">
