+++ src/app/constants/faqData.ts
@@ -1,0 +1,46 @@
+export const initialFaqData = [
+    {
+      question: "1. What services does Documitra offer?",
+      answer:
+        "Documitra provides comprehensive visa and passport assistance services, including document processing, application support, and expert guidance throughout your journey.",
+    },
+    {
+      question: "2. How does the Documitra process work?",
+      answer:
+        "Our streamlined process involves three simple steps: document submission, verification, and processing. Our experts guide you through each phase to ensure a smooth experience.",
+    },
+    {
+      question: "3. Is Documitra a government agency?",
+      answer:
+        "No, Documitra is an independent service provider that specializes in visa and passport assistance. We work alongside government agencies to facilitate your application process.",
+    },
+    {
+      question: "4. Is my personal information secure with Documitra?",
+      answer:
+        "Yes, we maintain the highest standards of data security. All personal information is encrypted and protected according to international privacy regulations.",
+    },
+    {
+      question:
+        "5. How long does it take to process a visa or passport application with Documitra?",
+      answer:
+        "Processing times vary depending on the type of application and the government's processing speed. We strive to expedite the process, but specific timelines may differ based on individual circumstances and government requirements.",
+    },
+  ];
+  
+  export const additionalFaqData = [
+    {
+      question: "6. What documents do I need to provide?",
+      answer:
+        "Required documents vary by application type. We'll provide a detailed checklist based on your specific needs and guide you through the collection process.",
+    },
+    {
+      question: "7. Do you offer expedited services?",
+      answer:
+        "Yes, we offer expedited processing for urgent applications, subject to availability and additional fees. Contact us to learn more about rush service options.",
+    },
+    {
+      question: "8. What countries do you service?",
+      answer:
+        "We assist with visa and passport applications for most major countries. Contact our team to confirm availability for your specific destination.",
+    },
+  ];
